services/telegram-bot/pyproject.toml
[project]
name = "bloger-telegram-bot"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.10",
    "httpx>=0.27",
    "pydantic-settings>=2.3",
    "structlog>=24.1",
    "pyyaml>=6.0",
    "sqlalchemy>=2.0",
    "asyncpg>=0.29",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bot"]
